+++ tests/pyen/catch_tab_error.py
@@ -1,7 +1,0 @@
-def test_tab_error(capsys):
-    from . import raise_tab_error
-
-    if not hasattr(raise_tab_error, "err"):
-        info = str(capsys.readouterr())
-        raise_tab_error.err = info
-    assert "TabError" in raise_tab_error.err